# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vessel API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Health",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Config",
        "description": "Client bootstrap: the default Bible version to use before the user picks one.",
        "operationId": "config",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                }
              }
            }
          }
        }
      }
    },
    "/api/bibles": {
      "get": {
        "tags": [
          "bible"
        ],
        "summary": "List Bibles",
        "operationId": "list_bibles",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "eng",
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bible"
                  },
                  "title": "Response List Bibles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bibles/{bible_id}/search": {
      "get": {
        "tags": [
          "bible"
        ],
        "summary": "Search Verses",
        "description": "Powers the `@` verse-lookup autocomplete in the journal editor.",
        "operationId": "search_verses",
        "parameters": [
          {
            "name": "bible_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bible Id"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Keyword or reference to search",
              "title": "Query"
            },
            "description": "Keyword or reference to search"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VerseHit"
                  },
                  "title": "Response Search Verses"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bibles/{bible_id}/verses/{verse_id}": {
      "get": {
        "tags": [
          "bible"
        ],
        "summary": "Get Verse",
        "operationId": "get_verse",
        "parameters": [
          {
            "name": "bible_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bible Id"
            }
          },
          {
            "name": "verse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Verse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/journals": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "List Journals",
        "operationId": "list_journals",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JournalRead"
                  },
                  "type": "array",
                  "title": "Response List Journals"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "journals"
        ],
        "summary": "Create Journal",
        "operationId": "create_journal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/journals/tags": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "List Tags",
        "description": "Distinct #idea tags across all entries, powering the `#` autocomplete.",
        "operationId": "list_tags",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  },
                  "type": "array",
                  "title": "Response List Tags"
                }
              }
            }
          }
        }
      }
    },
    "/api/journals/{journal_id}": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "Get Journal",
        "operationId": "get_journal",
        "parameters": [
          {
            "name": "journal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Journal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "journals"
        ],
        "summary": "Update Journal",
        "operationId": "update_journal",
        "parameters": [
          {
            "name": "journal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Journal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "journals"
        ],
        "summary": "Delete Journal",
        "operationId": "delete_journal",
        "parameters": [
          {
            "name": "journal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Journal Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Bible": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "abbreviation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbreviation"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Bible"
      },
      "Config": {
        "properties": {
          "default_bible_id": {
            "type": "string",
            "title": "Default Bible Id"
          }
        },
        "type": "object",
        "required": [
          "default_bible_id"
        ],
        "title": "Config"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Health": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "Health"
      },
      "JournalCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "entry_date": {
            "type": "string",
            "title": "Entry Date",
            "default": ""
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          }
        },
        "type": "object",
        "title": "JournalCreate"
      },
      "JournalRead": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "entry_date": {
            "type": "string",
            "title": "Entry Date",
            "default": ""
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "JournalRead"
      },
      "JournalUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "entry_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Date"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "title": "JournalUpdate"
      },
      "Tag": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "tag",
          "count"
        ],
        "title": "Tag"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verse": {
        "properties": {
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "reference",
          "content"
        ],
        "title": "Verse"
      },
      "VerseHit": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reference",
          "text"
        ],
        "title": "VerseHit"
      }
    }
  }
}